Fixed Listen2Stream Coverage
